api/requirements.txt
# Azure Functions v1 Programming Model - Static Web Apps Compatible
azure-functions==1.15.0
azure-functions-worker==1.0.0

# Core dependencies
python-dotenv==1.0.0
requests==2.31.0

# Build tools to prevent compilation issues
setuptools>=65.0.0
wheel>=0.38.0
